# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Nexus API
  description: API for the Nexus Investor & Entrepreneur Collaboration Platform
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local server
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [investor, entrepreneur]
      responses:
        '200':
          description: User registered successfully
  /auth/login:
    post:
      summary: Authenticate user & get token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
  /meetings:
    get:
      summary: Get user's meetings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of meetings
    post:
      summary: Schedule a meeting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                participant:
                  type: string
                title:
                  type: string
                date:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Meeting scheduled
  /documents:
    get:
      summary: Get all documents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of documents
    post:
      summary: Upload a document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                  format: binary
                title:
                  type: string
      responses:
        '200':
          description: Document uploaded
  /payments/deposit:
    post:
      summary: Mock deposit funds
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
      responses:
        '200':
          description: Transaction completed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT